% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_data.R
\name{aggregate_raw_data}
\alias{aggregate_raw_data}
\title{Aggregate raw data}
\usage{
aggregate_raw_data(isofiles, gather = FALSE)
}
\arguments{
\item{isofiles}{collection of isofile objects}

\item{gather}{whether to gather data into long format after aggregation (e.g. for plotting)}
}
\description{
Aggregate raw data
}
\seealso{
Other data aggregation functions: \code{\link{aggregate_file_info}},
  \code{\link{aggregate_standards_info}},
  \code{\link{aggregate_vendor_data_table}}
}
