% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_data.R
\name{aggregate_vendor_data_table}
\alias{aggregate_vendor_data_table}
\title{Aggregate vendor computed table data}
\usage{
aggregate_vendor_data_table(isofiles, with_units = TRUE)
}
\arguments{
\item{isofiles}{collection of isofile objects}

\item{with_units}{whether to include units in the column headers or not}
}
\description{
Aggregate vendor computed table data
}
\seealso{
Other data aggregation functions: \code{\link{aggregate_file_info}},
  \code{\link{aggregate_raw_data}},
  \code{\link{aggregate_standards_info}}
}
