% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{caching}
\alias{caching}
\alias{turn_caching_on}
\alias{turn_caching_off}
\title{Turn caching on/off}
\usage{
turn_caching_on(data)

turn_caching_off(data)
}
\arguments{
\item{data}{a data frame - returned as is if provided (e.g. in the middle of a pipeline)}
}
\description{
These functions turn caching of data files (and reading from cache) on/off in all subsequent isoread calls by changing the global settings for the \code{cache} parameter. Can be called stand alone or within a pipeline.
}
