% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cleanup_isoreader_cache}
\alias{cleanup_isoreader_cache}
\title{Cleanup old cached files}
\usage{
cleanup_isoreader_cache(all = FALSE)
}
\arguments{
\item{all}{if set to TRUE, all cached files will be removed regardless of their version}
}
\description{
Removes old cached files generated by isoreader versions different from the currently installed package. Removes all cached files if \code{clean_all} is set to TRUE.
}
