% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_conversion.R
\name{convert_time}
\alias{convert_time}
\title{Convert time units in continuous flow files}
\usage{
convert_time(isofiles, to, quiet = setting("quiet"))
}
\arguments{
\item{isofiles}{collection of isofile objects}

\item{to}{what time units to convert to}

\item{quiet}{whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages. Set parameter to overwrite global defaults for this function or set global defaults with calls to \link[=info_messages]{turn_info_message_on} and \link[=info_messages]{turn_info_message_off}}
}
\value{
the passed in isofile(s) with changed time units
}
\description{
This function can be used to convert the time units of a collection of isofiles to a new common time unit. The original time units are inferred from the naming of the time column. New time units are very flexible and can be anything that \code{\link[lubridate]{duration}} understands, i.e. "s", "seconds", "min", "minutes", "hours", "days", etc. are all valid units.
}
