% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_scaling.R
\name{convert_voltages_to_currents}
\alias{convert_voltages_to_currents}
\title{Convert voltages to currents
Automatically finds the matching voltage columns that fit the resistor mass names and converts to currents in the supplied I_units. All voltage columns should be in units of V_units.}
\usage{
convert_voltages_to_currents(data, R, V_units = "mV", I_units = "nA",
  R_units = "GOhm", V_pattern = "^[vV](\\\\d+)", I_prefix = "i",
  quiet = setting("quiet"))
}
\arguments{
\item{data}{the data frame}

\item{R}{resistor value(s), named vector with Rm (m=mass) as names (e.g. c(R45=0.3, R46=3)), in units of R_units}

\item{V_units}{what units voltages are in (default mV, i.e. 10^-3)}

\item{I_units}{what units currents are in (default nA, i.e. 10^-9)}

\item{R_units}{what units resistances are in (default GOhm, i.e. 10^9)}

\item{V_pattern}{regular expression pattern how to recognize voltage columns and detect the masses they belong to (default is v followed by a number, e.g. v45)}

\item{I_prefix}{prefix for the newly created current columns, the suffix is automatically the current units}

\item{quiet}{whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages. Set parameter to overwrite global defaults for this function or set global defaults with calls to \link[=info_messages]{turn_info_message_on} and \link[=info_messages]{turn_info_message_off}}
}
\description{
Convert voltages to currents
Automatically finds the matching voltage columns that fit the resistor mass names and converts to currents in the supplied I_units. All voltage columns should be in units of V_units.
}
\note{
make sure that existing currents are not overwritten (e.g. in data frame that comes jointly from isodat and elementar)

consider automatically guessing V_units from the voltage columns?
}
