% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isodata_structures.R
\name{is_isofile}
\alias{is_isofile}
\alias{is_isofile_list}
\alias{is_iso_object}
\title{Isoreader data structure functions}
\usage{
is_isofile(x)

is_isofile_list(x)

is_iso_object(x)
}
\arguments{
\item{x}{An object}
}
\description{
\code{is_isofile} tests if the object is an isofile

\code{is_isofile_list} tests if the object is an isofile list (collection of isofiles)

\code{is_iso_object} test if the object is an iso-object (isofile or isofile list)
}
