% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_data.R, R/export.R, R/package.R,
%   R/problems.R
\name{iso_get_data}
\alias{iso_get_data}
\alias{iso_get_standards_info}
\alias{iso_get_resistors_info}
\alias{iso_export_to_excel}
\alias{iso_export_to_feather}
\alias{deprecated}
\alias{iso_omit_files_with_problems}
\title{These functions have been deprecated and will be removed in future versions of isoreader.}
\usage{
iso_get_data(...)

iso_get_standards_info(...)

iso_get_resistors_info(...)

iso_export_to_excel(...)

iso_export_to_feather(...)

iso_omit_files_with_problems(...)
}
\arguments{
\item{...}{deprecated (passed on to new function if superseded)}
}
\description{
These functions have been deprecated and will be removed in future versions of isoreader.
}
\details{
\code{iso_get_data}: use \link{iso_get_all_data} instead

\code{iso_get_standards_info}: use \link{iso_get_standards} instead

\code{iso_get_resistors_info}: use \link{iso_get_resistors} instead

\code{iso_export_to_excel}: use \link{iso_export_files_to_excel} instead

\code{iso_export_to_feather}: use \link{iso_export_files_to_feather} instead

\code{iso_omit_files_with_problems}: use \link{iso_filter_files_with_problems} instead
}
