% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_to_excel}
\alias{export_to_excel}
\title{Export data to Excel}
\usage{
export_to_excel(isofiles, filepath, include_raw_data = TRUE,
  include_file_info = TRUE, include_method_info = TRUE,
  include_vendor_data_table = TRUE, include_problems = TRUE,
  quiet = default(quiet))
}
\arguments{
\item{isofiles}{collection of isofile objects}

\item{filepath}{the path (folder and filename) to the export file. The correct file extension is automatically added if not already in the filename, i.e. filename can be provided with or without extension.}

\item{include_raw_data}{whether to include the raw data in the export (if available)}

\item{include_file_info}{whether to include the file info in the export (if available)}

\item{include_method_info}{whether to include methods infor in the export (if available)}

\item{include_vendor_data_table}{whether to include the vendor data table in the export (if available)}

\item{include_problems}{whether to include the problems table}

\item{quiet}{whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages. Set parameter to overwrite global defaults for this function or set global defaults with calls to \link[=info_messages]{turn_info_message_on} and \link[=info_messages]{turn_info_message_off}}
}
\value{
returns the isofiles object invisibly for use in pipelines
}
\description{
This function exports the passed in isofiles to Excel. The different kinds of data (raw data, file info, methods info, etc.) are exported to separate tabs within the excel file but they are only exported if the corresponding \code{include_} parameter is set to \code{TRUE} and only for data types for which this type of data is available and was read (see \code{\link{read_dual_inlet}}, \code{\link{read_continuous_flow}} for details on read parameters).
}
\seealso{
Other export functions: \code{\link{export_to_feather}},
  \code{\link{export_to_rda}}
}
