% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_to_feather}
\alias{export_to_feather}
\title{Export to feather}
\usage{
export_to_feather(isofiles, filepath_prefix, zip = FALSE,
  include_raw_data = TRUE, include_file_info = TRUE,
  include_method_info = TRUE, include_vendor_data_table = TRUE,
  quiet = setting("quiet"))
}
\arguments{
\item{isofiles}{collection of isofile objects}

\item{filepath_prefix}{the path (folder and filename) prefix for the exported feather files. The correct suffix for different kinds of data and file extension is automatically added}

\item{include_raw_data}{whether to include the raw data in the export (if available)}

\item{include_file_info}{whether to include the file info in the export (if available)}

\item{include_vendor_data_table}{whether to include the vendor data table in the export (if available)}

\item{quiet}{whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages. Set parameter to overwrite global defaults for this function or set global defaults with calls to \link[=info_messages]{turn_info_message_on} and \link[=info_messages]{turn_info_message_off}}
}
\value{
returns the isofiles object invisibly for use in pipelines
}
\description{
This function exports the passed in isofiles to the Python and R shared feather file format. The different kinds of data (raw data, file info, methods info, etc.) are exported to separate feather files that are saved with the provided \code{filepath_prefix} as prefix. All are only exported if the corresponding \code{include_} parameter is set to \code{TRUE} and only for data types for which this type of data is available and was read (see \code{\link{read_dual_inlet}}, \code{\link{read_continuous_flow}} for details on read parameters).
}
\seealso{
Other export functions: \code{\link{export_to_excel}},
  \code{\link{export_to_rda}}
}
