% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_data.R
\name{get_file_id}
\alias{get_file_id}
\alias{get_file_info}
\alias{get_file_path}
\alias{get_file_subpath}
\alias{get_file_datetime}
\title{Get file information}
\usage{
get_file_id(isofile)

get_file_path(isofile)

get_file_subpath(isofile)

get_file_datetime(isofile)
}
\arguments{
\item{isofile}{an isofile to retrieve basic file information from}
}
\description{
Retrieve basic file information form an individual isotope file (isofile) object. All of these can also be recoverd for an entire set of files using \code{\link{aggregate_file_info}} and specifiying which info to recover, for example, \code{include = c("file_id", "file_path", "file_datetime")}
}
\details{
\code{get_file_id()}: retrieve the file ID (this is typially the file name)

\code{get_file_path()}: retrieve the file path (this is the path to the file in case of single file formats such as .dxf or .did and the path to the archieve file in case of collective file formats such as .iarc)

\code{get_file_subpath()}: retrieve the file subpath (this only exists for collective file formats such as .iarc and is the name of the metadata file inside the .iarc archive). Returns NA for isofile without subpath.

\code{get_file_datetime()}: retrieve the run date and time in \code{\link[base]{POSIXct}} format
}
\seealso{
Other data retrieval functions: \code{\link{aggregate_file_info}},
  \code{\link{aggregate_raw_data}},
  \code{\link{aggregate_resistors_info}},
  \code{\link{aggregate_standards_info}},
  \code{\link{aggregate_vendor_data_table}}
}
