% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filter_files}
\alias{filter_files}
\title{Filter isofiles}
\usage{
filter_files(isofiles, ..., quiet = setting("quiet"))
}
\arguments{
\item{isofiles}{collection of isofile objects}

\item{...}{filter conditions applied based on each file's file_info (see \code{\link{aggregate_file_info}})}

\item{quiet}{whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages. Set parameter to overwrite global defaults for this function or set global defaults with calls to \link[=info_messages]{turn_info_message_on} and \link[=info_messages]{turn_info_message_off}}
}
\description{
Filter for specific isofiles using file info parameters, most commonlty file_id and file_datetime but others can be used as well if they exist.
}
