% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_binary_isodat_files.R
\name{find_isodat_structure_blocks}
\alias{find_isodat_structure_blocks}
\title{find all isodat structure blocks}
\usage{
find_isodat_structure_blocks(bfile, unknown_block_n_chars = 8L)
}
\arguments{
\item{bfile}{the isodat binary file object (must have $raw set)}

\item{unknown_block_n_chars}{the number of chars to preview as 'block' text in the resulting tibble}
}
\description{
find all isodat structure blocks
}
