% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_binary_isodat_files.R
\name{find_pattern_blocks}
\alias{find_pattern_blocks}
\title{find regular expression pattern and turn into a block tibble}
\usage{
find_pattern_blocks(
  raw,
  regex,
  start_expr,
  len_expr,
  data_len_expr,
  block_expr
)
}
\arguments{
\item{raw}{binary vector}

\item{start_expr}{expression to calculate starting point (relative to the regexp match var 'pos')}

\item{len_expr}{expression to calculate the length of the block}

\item{data_len_expr}{expression to calculate the length of the data in the block}

\item{block_expr}{expression to construct the block text}

\item{regular}{expression to match}
}
\description{
find regular expression pattern and turn into a block tibble
}
