% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_binary_isodat_files.R
\name{find_unknown_blocks}
\alias{find_unknown_blocks}
\title{find unknown patterns and turn into a block tibble}
\usage{
find_unknown_blocks(raw, blocks)
}
\arguments{
\item{raw}{binary vector}

\item{blocks}{tibble with identified blocks, must have columns start & end}
}
\description{
find unknown patterns and turn into a block tibble
}
