% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_binary_isodat_files.R
\name{format_isodat_structure_blocks}
\alias{format_isodat_structure_blocks}
\title{format isodat structure blocks for printout}
\usage{
format_isodat_structure_blocks(
  bfile,
  new_line_blocks = type \%in\% c("C block", "x-000"),
  indent_blocks = type == "x-000",
  unknown_block_n_chars = 8L,
  data_blocks = type \%in\% c("text", "unknown"),
  data_highlight = FALSE,
  pos_info = TRUE
)
}
\arguments{
\item{bfile}{the isodat binary file object (must have $raw set)}

\item{new_line_blocks}{expression when to create a new line}

\item{indent_blocks}{expression when to indent a line (only if also matched by new_line_blocks)}

\item{unknown_block_n_chars}{the number of chars to preview as 'block' text in the resulting tibble}

\item{data_blocks}{expression to mark data blocks}

\item{data_highlight}{expression to insert a 'HIGHLIGHT' marker in the text, example `len > 1000` to highlight large data blocks}

\item{pos_info}{whether to include position information for each line  (highly recommended)}
}
\description{
format isodat structure blocks for printout
}
