% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_data.R
\name{get_data_table}
\alias{get_data_table}
\title{Aggregate table data}
\usage{
get_data_table(isofiles)
}
\arguments{
\item{isofiles}{isotope file objects}
}
\description{
Aggregate table data
}
\seealso{
Other data aggregation functions: \code{\link{get_file_info}},
  \code{\link{get_raw_data}}
}
