% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_data.R
\name{get_file_info}
\alias{get_file_info}
\title{Aggregate file info}
\usage{
get_file_info(isofiles)
}
\arguments{
\item{isofiles}{collection of isofile objects}
}
\description{
Note file info entries with multiple values are concatenated for this general purpose function.
To get access to a specific multi-value file info, access using $file_info[['INFO_NAME']]
}
\seealso{
Other data aggregation functions: \code{\link{get_method_info_standards}},
  \code{\link{get_raw_data}},
  \code{\link{get_vendor_data_table}}
}
