% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_binary_isodat_files.R
\name{get_isodat_control_blocks_config}
\alias{get_isodat_control_blocks_config}
\title{returns a tibble with control blacks for isodat files}
\usage{
get_isodat_control_blocks_config()
}
\value{
tibble with control blocks for isodat
}
\description{
returns a tibble with control blacks for isodat files
}
