% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_data.R
\name{get_isofile_id}
\alias{get_isofile_id}
\alias{get_isofile_path}
\alias{get_isofile_subpath}
\alias{get_isofile_datetime}
\title{Get file information}
\usage{
get_isofile_id(isofile)

get_isofile_path(isofile)

get_isofile_subpath(isofile)

get_isofile_datetime(isofile)
}
\arguments{
\item{isofile}{an isofile to retrieve basic file information from}
}
\description{
Retrieve basic file information form an isotope file (isofile) object.
}
\details{
\code{get_isofile_id()}: retrieve the file ID (this is typially the file name)

\code{get_isofile_path()}: retrieve the file path (this is the path to the file in case of single file formats such as .dxf or .did and the path to the archieve file in case of collective file formats such as .iarc)

\code{get_isofile_subpath()}: retrieve the file subpath (this only exists for collective file formats such as .iarc and is the name of the metadata file inside the .iarc archive). Returns NA for isofile without subpath.

\code{get_isofile_datetime()}: retrieve the run date and time
}
