% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_data.R
\name{get_mass_data}
\alias{get_mass_data}
\title{Aggregate mass data}
\usage{
get_mass_data(isofiles)
}
\arguments{
\item{isofiles}{isotope file objects}
}
\description{
Aggregate mass data
}
\seealso{
Other data_aggregation: \code{\link{get_data_table}},
  \code{\link{get_file_info}}
}
