% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_data.R
\name{get_raw_data}
\alias{get_raw_data}
\title{Aggregate raw data}
\usage{
get_raw_data(isofiles)
}
\arguments{
\item{isofiles}{isotope file objects}
}
\description{
Aggregate raw data
}
\seealso{
Other data aggregation functions: \code{\link{get_file_info}},
  \code{\link{get_vendor_computed_data_table}}
}
