% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_prefix_scaling.R
\name{get_si_prefix_scaling}
\alias{get_si_prefix_scaling}
\title{get the scaling factor for a simple si unit prefix
does not currently process compound units (e.g. m/s) or powers (e.g. cm2 s-1)}
\usage{
get_si_prefix_scaling(unit, suffix)
}
\arguments{
\item{unit}{the unit to find the suffix for (e.g. mm, kg, ms, nA, kV)}

\item{suffix}{the expected suffix (e.g. m, g, s, A, V)}
}
\description{
get the scaling factor for a simple si unit prefix
does not currently process compound units (e.g. m/s) or powers (e.g. cm2 s-1)
}
