% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_binary_isodat_files.R
\name{get_unknown_blocks_text}
\alias{get_unknown_blocks_text}
\title{updates block information for unknown blocks}
\usage{
get_unknown_blocks_text(blocks, raw, unknown_block_n_chars = 8L)
}
\description{
updates block information for unknown blocks
}
