% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{info_messages}
\alias{info_messages}
\alias{turn_info_messages_on}
\alias{turn_info_messages_off}
\title{Turn information messages on/off}
\usage{
turn_info_messages_on(data)

turn_info_messages_off(data)
}
\arguments{
\item{data}{a data frame - returned invisibly as is if provided (e.g. in the middle of a pipeline)}
}
\description{
These functions turn information messages on/off in all subsequent function calls by changing the global settings for the \code{quiet} parameter of most isoreader functions. These functions can be called stand alone or within a pipeline to turn messages on/off at a certain point during the pipeline.
}
\seealso{
Other settings functions: \code{\link{caching}},
  \code{\link{set_default_read_parameters}},
  \code{\link{show_isoreader_parameters}}
}
