% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isodata_structures.R
\name{is_isofile}
\alias{is_isofile}
\title{Test if the object is an isofile}
\usage{
is_isofile(x)
}
\arguments{
\item{x}{An object}
}
\value{
`TRUE` if the object inherits from the `isofile` class.
}
\description{
Test if the object is an isofile
}
