% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isodata_structures.R
\name{is_isofile_list}
\alias{is_isofile_list}
\title{Test if the object is an isofile list (collection of isofiles)}
\usage{
is_isofile_list(x)
}
\arguments{
\item{x}{An object}
}
\value{
`TRUE` if the object inherits from the `isofile_list` class.
}
\description{
Test if the object is an isofile list (collection of isofiles)
}
