% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{iso_caching}
\alias{iso_caching}
\alias{iso_turn_reader_caching_on}
\alias{iso_turn_reader_caching_off}
\title{Turn caching on/off}
\usage{
iso_turn_reader_caching_on(data = NULL)

iso_turn_reader_caching_off(data = NULL)
}
\arguments{
\item{data}{a data frame - returned invisibly as is if provided (e.g. in the middle of a pipeline)}
}
\description{
These functions turn caching of data files (and reading from cache) on/off in all subsequent isoread calls by changing the global settings for the \code{cache} parameter. Can be called stand alone or within a pipeline.
}
\seealso{
Other settings functions: 
\code{\link{iso_get_default_reader_parameters}()},
\code{\link{iso_info_messages}},
\code{\link{iso_set_default_read_parameters}()}
}
\concept{settings functions}
