% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{iso_cleanup_reader_cache}
\alias{iso_cleanup_reader_cache}
\title{Cleanup old cached files}
\usage{
iso_cleanup_reader_cache(all = FALSE)
}
\arguments{
\item{all}{if set to TRUE, all cached files will be removed regardless of their version}
}
\description{
Removes old cached files generated by isoreader versions different from the currently installed package. Removes all cached files if \code{clean_all} is set to TRUE.
}
