% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{iso_debug_mode}
\alias{iso_debug_mode}
\alias{iso_turn_debug_on}
\alias{iso_turn_debug_off}
\alias{set_read_file_event_expr}
\alias{set_finish_file_event_expr}
\title{Debugging functions}
\usage{
iso_turn_debug_on(data = NULL, catch_errors = TRUE, cache = FALSE)

iso_turn_debug_off(data = NULL)

set_read_file_event_expr(event_expr = NULL)

set_finish_file_event_expr(event_expr = NULL)
}
\arguments{
\item{data}{a data frame - returned invisibly as is if provided (e.g. in the middle of a pipeline)}

\item{catch_errors}{whether to still catch errors in debug mode or whether to throw them}

\item{cache}{whether to cache or read anything from cache}

\item{event_expr}{an expression to evaluate in the context of reading individual iso files (evaluated in the local environment at the beginning of a file read)}
}
\description{
For troubleshooting. Not exported.
}
