% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{iso_export_to_excel}
\alias{iso_export_to_excel}
\title{Function renamed}
\usage{
iso_export_to_excel(...)
}
\arguments{
\item{...}{deprecated}
}
\description{
Use \link{iso_export_files_to_excel} instead.
}
