% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{iso_export_to_feather}
\alias{iso_export_to_feather}
\title{Function renamed}
\usage{
iso_export_to_feather(...)
}
\arguments{
\item{...}{deprecated}
}
\description{
Use \link{iso_export_files_to_feather} instead.
}
