% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{iso_export_to_rda}
\alias{iso_export_to_rda}
\title{Export data to R Data Archive (.rda) (deprecated)}
\usage{
iso_export_to_rda(iso_files, filepath, quiet = default(quiet))
}
\arguments{
\item{iso_files}{collection of iso_file objects}

\item{filepath}{the path (folder and filename) to the export file. The correct file extension is automatically added if not already in the filename, i.e. filename can be provided with or without extension.}

\item{quiet}{whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages. Set parameter to overwrite global defaults for this function or set global defaults with calls to \link[=iso_info_messages]{iso_turn_info_message_on} and \link[=iso_info_messages]{iso_turn_info_message_off}}
}
\value{
returns the iso_files object invisibly for use in pipelines
}
\description{
This function is deprecated. Please use \code{\link{iso_save}} instead to save collections of isofiles.
}
\seealso{
Other export functions: 
\code{\link{iso_export_to_excel}()},
\code{\link{iso_export_to_feather}()},
\code{\link{iso_save}()}
}
\concept{export functions}
