% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{iso_filter_files}
\alias{iso_filter_files}
\title{Filter isofiles}
\usage{
iso_filter_files(isofiles, ..., quiet = default(quiet))
}
\arguments{
\item{isofiles}{collection of isofile objects}

\item{...}{filter conditions applied based on each file's file_info (see \code{\link{iso_aggregate_file_info}})}

\item{quiet}{whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages. Set parameter to overwrite global defaults for this function or set global defaults with calls to \link[=iso_info_messages]{turn_info_message_on} and \link[=iso_info_messages]{turn_info_message_off}}
}
\description{
Filter for specific isofiles using file info parameters, most commonlty file_id and file_datetime but others can be used as well if they exist.
}
