% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_data.R
\name{iso_get_data_summary}
\alias{iso_get_data_summary}
\title{Get data summary}
\usage{
iso_get_data_summary(iso_files, quiet = default(quiet))
}
\arguments{
\item{iso_files}{single iso file or collection of iso_file objects}
}
\value{
a \code{\link[tibble]{data_frame}} that summarizes the data in the \code{iso_files}
}
\description{
Summarize the data information from one or multiple iso files.
}
