% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{iso_get_default_reader_parameters}
\alias{iso_get_default_reader_parameters}
\title{Get the current default parameters}
\usage{
iso_get_default_reader_parameters()
}
\description{
Retrieve a table with all default function parameters for this package. 
To set reader parameters, see \code{\link{iso_set_default_read_parameters}}.
For a piping compatible version of this function, see \link{iso_show_default_reader_parameters}.
}
\seealso{
Other settings functions: \code{\link{iso_caching}},
  \code{\link{iso_info_messages}},
  \code{\link{iso_set_default_read_parameters}},
  \code{\link{iso_show_default_reader_parameters}}
}
