% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_data.R
\name{iso_get_file_info}
\alias{iso_get_file_info}
\title{Aggregate file info}
\usage{
iso_get_file_info(iso_files, select = everything(),
  quiet = default(quiet))
}
\arguments{
\item{iso_files}{collection of iso_file objects}

\item{select}{which columns to select - use \code{c(...)} to select multiple, supports all \link[dplyr]{select} syntax including renaming columns. File id is always included and cannot be renamed.}

\item{quiet}{whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages. Set parameter to overwrite global defaults for this function or set global defaults with calls to \link[=iso_info_messages]{iso_turn_info_message_on} and \link[=iso_info_messages]{iso_turn_info_message_off}}
}
\description{
Combine file information from multiple iso_files. By default all information is included but specific columns can be targeted using the \code{select} parameter, which uses the \link{\code{iso_select_file_info}} function to select and/or rename columns. File information beyond \code{file_id}, \code{file_root}, \code{file_path} and \code{file_datetime} is only available if the \code{iso_files} were read with parameter \code{read_file_info=TRUE}.
}
\note{
File info entries with multiple values remain nested multi-value (=list) columns and can be unnested using \link[tidyr]{unnest}.
}
\seealso{
Other data retrieval functions: \code{\link{iso_get_bgrd_data}},
  \code{\link{iso_get_data}},
  \code{\link{iso_get_raw_data}},
  \code{\link{iso_get_resistors_info}},
  \code{\link{iso_get_standards_info}},
  \code{\link{iso_get_vendor_data_table}}
}
\concept{data retrieval functions}
