% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problems.R
\name{iso_get_problems}
\alias{iso_get_problems}
\title{Retrieve parsing problems}
\usage{
iso_get_problems(x)
}
\description{
This is a re-export of the readr \code{\link[readr]{problems}} function.
}
\seealso{
Other problem functions: \code{\link{iso_get_problems_summary}},
  \code{\link{iso_omit_files_with_problems}},
  \code{\link{iso_problem_functions}}
}
