% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problems.R
\name{iso_get_problems_summary}
\alias{iso_get_problems_summary}
\title{Retrieve a summary of the problems}
\usage{
iso_get_problems_summary(iso_files, problem_files_only = TRUE)
}
\arguments{
\item{iso_files}{collection of iso_file objects}

\item{problem_files_only}{whether to list only problem files or all files}
}
\value{
data frame with file_id and number of encountered errors and warnings
}
\description{
Returns a data frame listing how many errors and warnings were encountered for each file. For details on each error/warning, see \link[readr]{problems} and the \link{iso_problem_functions}.
}
\seealso{
Other problem functions: \code{\link{iso_get_problems}},
  \code{\link{iso_has_problems}},
  \code{\link{iso_omit_files_with_problems}},
  \code{\link{iso_problem_functions}}
}
