% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_data.R
\name{iso_get_raw_data}
\alias{iso_get_raw_data}
\title{Aggregate raw data}
\usage{
iso_get_raw_data(iso_files, gather = FALSE, include_file_info = c(),
  quiet = default(quiet))
}
\arguments{
\item{iso_files}{collection of iso_file objects}

\item{gather}{whether to gather data into long format after aggregation (e.g. for plotting)}

\item{include_file_info}{if provided, will include the requested file information (see \code{\link{iso_get_file_info}}) with the raw data}

\item{quiet}{whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages. Set parameter to overwrite global defaults for this function or set global defaults with calls to \link[=iso_info_messages]{turn_info_message_on} and \link[=iso_info_messages]{turn_info_message_off}}
}
\description{
Aggregate the raw ion data from the provided iso_files. Can aggregate either in a wide table (for easy overview) or a gathered long table (for plotting and further data processing). The raw data is only available if the iso_files were read with parameter \code{read_raw_data=TRUE}.
}
\seealso{
Other data retrieval functions: \code{\link{iso_get_file_info}},
  \code{\link{iso_get_resistors_info}},
  \code{\link{iso_get_standards_info}},
  \code{\link{iso_get_vendor_data_table}}
}
