% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_data.R
\name{iso_get_standards_info}
\alias{iso_get_standards_info}
\title{Aggregate standards from methods info}
\usage{
iso_get_standards_info(iso_files, with_ratios = FALSE,
  include_file_info = NULL, quiet = default(quiet))
}
\arguments{
\item{iso_files}{collection of iso_file objects}

\item{with_ratios}{whether to include ratios or just standard delta values}

\item{include_file_info}{which file information to include (see \code{\link{iso_get_file_info}}). Use \code{c(...)} to select multiple, supports all \link[dplyr]{select} syntax including renaming columns.}

\item{quiet}{whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages. Set parameter to overwrite global defaults for this function or set global defaults with calls to \link[=iso_info_messages]{iso_turn_info_message_on} and \link[=iso_info_messages]{iso_turn_info_message_off}}
}
\description{
Aggregates the isotopic standard information recovered from the provided iso_files. Can aggregate just the standards' delta values or combine the delta values with the recovered ratios (if any). Use paramter \code{with_ratios} to exclude/include the ratios. This information is only available if the iso_files were read with parameter \code{read_method_info=TRUE}.
}
\seealso{
Other data retrieval functions: \code{\link{iso_get_data}},
  \code{\link{iso_get_file_info}},
  \code{\link{iso_get_raw_data}},
  \code{\link{iso_get_resistors_info}},
  \code{\link{iso_get_vendor_data_table}}
}
