% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{iso_get_supported_file_types}
\alias{iso_get_supported_file_types}
\title{Supported file types}
\usage{
iso_get_supported_file_types()
}
\description{
Get an overview of all the file types currently supported by the isoreader package.
}
