% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{iso_get_units}
\alias{iso_get_units}
\title{Retrieve number units}
\usage{
iso_get_units(x)
}
\arguments{
\item{x}{variable to get the units for (vector or data frame)}
}
\description{
This function returns the units of a numerical value generated by \code{\link{iso_double_with_units}}. It returns \code{NA}) for unitless variables. Returns a column-named vector of units if \code{x} is a data frame / tibble. Returns the direct units of \code{x} in all other cases.
}
\seealso{
Other functions for values with units: 
\code{\link{iso_is_double_with_units}()},
\code{\link{iso_make_units_explicit}()},
\code{\link{iso_make_units_implicit}()},
\code{\link{iso_strip_units}()},
\code{\link{iso_with_units}()}
}
\concept{functions for values with units}
