% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problems.R
\name{iso_has_problems}
\alias{iso_has_problems}
\title{Check for parsing problems}
\usage{
iso_has_problems(iso_files)
}
\arguments{
\item{iso_files}{collection of iso_file objects}
}
\value{
boolean
}
\description{
Check for parsing problems
}
\seealso{
Other problem functions: 
\code{\link{iso_filter_files_with_problems}()},
\code{\link{iso_get_problems_summary}()},
\code{\link{iso_get_problems}()},
\code{\link{iso_problem_functions}}
}
\concept{problem functions}
