% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{iso_info_messages}
\alias{iso_info_messages}
\alias{iso_turn_info_messages_on}
\alias{iso_turn_info_messages_off}
\title{Turn information messages on/off}
\usage{
iso_turn_info_messages_on(data = NULL)

iso_turn_info_messages_off(data = NULL)
}
\arguments{
\item{data}{a data frame - returned invisibly as is if provided (e.g. in the middle of a pipeline)}
}
\description{
These functions turn information messages on/off in all subsequent function calls by changing the global settings for the \code{quiet} parameter of most isoreader functions. These functions can be called stand alone or within a pipeline to turn messages on/off at a certain point during the pipeline.
}
\seealso{
Other settings functions: \code{\link{iso_caching}},
  \code{\link{iso_get_default_reader_parameters}},
  \code{\link{iso_set_default_read_parameters}},
  \code{\link{iso_show_default_reader_parameters}}
}
