% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{iso_is_double_with_units}
\alias{iso_is_double_with_units}
\title{Check if a value has units}
\usage{
iso_is_double_with_units(x)
}
\arguments{
\item{x}{vector to check for whether it is a double with units}
}
\description{
Check if a variable is a double with units. That is if it has been generated by \code{\link{iso_double_with_units}}.
}
\seealso{
Other functions for values with units: 
\code{\link{iso_get_units}()},
\code{\link{iso_make_units_explicit}()},
\code{\link{iso_make_units_implicit}()},
\code{\link{iso_strip_units}()},
\code{\link{iso_with_units}()}
}
\concept{functions for values with units}
