% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problems.R
\name{iso_omit_files_with_problems}
\alias{iso_omit_files_with_problems}
\title{Remove problematic files}
\usage{
iso_omit_files_with_problems(iso_files, type = c("error", "warning", "both"),
  quiet = default(quiet))
}
\arguments{
\item{iso_files}{collection of iso_file objects}

\item{type}{what type of problem causes removal of the file: \code{"error"}, \code{"warning"} or \code{"both"}}

\item{quiet}{whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages. Set parameter to overwrite global defaults for this function or set global defaults with calls to \link[=iso_info_messages]{turn_info_message_on} and \link[=iso_info_messages]{turn_info_message_off}}
}
\description{
Removes the files that have encountered problems, either errors, warnings or both and returns the remaining iso_files. For additional functions available to check for and deal with problems, see the \link{problem_functions}.
}
\seealso{
Other problem functions: \code{\link{iso_problems_summary}},
  \code{\link{iso_problems}},
  \code{\link{problem_functions}}
}
