% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{iso_plot_raw_data}
\alias{iso_plot_raw_data}
\title{Plot raw data from isoreader files}
\usage{
iso_plot_raw_data(iso_files, ..., quiet = default(quiet))
}
\arguments{
\item{iso_files}{collection of iso_file objects}

\item{...}{parameters for the data specific plotting functions}

\item{quiet}{whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages. Set parameter to overwrite global defaults for this function or set global defaults with calls to \link[=iso_info_messages]{turn_info_message_on} and \link[=iso_info_messages]{turn_info_message_off}}
}
\description{
Convenience function for making standard plots for raw isoreader data. Calls \code{\link{iso_plot_continuous_flow}} and \code{\link{iso_plot_dual_inlet}} for data specific plotting (see those functions for parameter details).
}
\seealso{
Other plot functions: \code{\link{iso_plot_continuous_flow}}
}
