% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_binary_isodat_files.R
\name{iso_print_source_file_structure}
\alias{iso_print_source_file_structure}
\alias{iso_print_source_file_structure.binary_isodat_file}
\title{Print source file structure}
\usage{
iso_print_source_file_structure(object, ..., save_to_file = NULL)

\method{iso_print_source_file_structure}{binary_isodat_file}(
  object,
  start = 1,
  length = NULL,
  end = start + length,
  ...,
  save_to_file = NULL
)
}
\arguments{
\item{object}{the object for which to print the source file structure.}

\item{...}{additional parameters depending on source file types}

\item{save_to_file}{whether to save the source file structure to a text file (provide file path, will overwrite anything already in the file!)}

\item{start}{starting position in the binary file to print from (prints the first block that spans this range)}

\item{length}{length in the binary file to print to (by default \code{NULL}, which means print everything)}

\item{end}{until which point in the binary file to print to. If provided, overrides whatever is specified in \code{length}}
}
\value{
the source file structure (invisibly if it is also saved to a file via \code{save_to_file})
}
\description{
Debugging function to print a representation of the structure of the source file underlying an isofile (if there is one).
}
