% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isodata_structures.R
\name{print.iso_file_list}
\alias{print.iso_file_list}
\alias{print.iso_file}
\alias{print.dual_inlet}
\alias{print.continuous_flow}
\alias{print.scan}
\title{Isofile printing}
\usage{
\method{print}{iso_file_list}(x, ...)

\method{print}{iso_file}(x, ..., show_problems = TRUE)

\method{print}{dual_inlet}(x, ..., show_problems = TRUE)

\method{print}{continuous_flow}(x, ..., show_problems = TRUE)

\method{print}{scan}(x, ..., show_problems = TRUE)
}
\arguments{
\item{x}{Object to show.}

\item{...}{additional parameters passed to print.default}

\item{show_problems}{whether to show encountered problems}
}
\description{
Print summary of individual iso_files (dual inlet or continuous flow) or collection of iso_files.
}
