% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problems.R
\name{iso_problems_summary}
\alias{iso_problems_summary}
\title{Retrieve a summary of the problems}
\usage{
iso_problems_summary(iso_files)
}
\arguments{
\item{iso_files}{collection of iso_file objects}
}
\value{
data frame with file_id and number of encountered errors and warnings
}
\description{
Returns a data frame listing how many errors and warnings were encountered for each file. For details on each error/warning, see \link[readr]{problems} and the \link{problem_functions}.
}
\seealso{
Other problem functions: \code{\link{iso_omit_files_with_problems}},
  \code{\link{iso_problem_functions}},
  \code{\link{iso_problems}}
}
