% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoread.R
\name{iso_read_continuous_flow}
\alias{iso_read_continuous_flow}
\title{Load continuous flow data}
\usage{
iso_read_continuous_flow(
  ...,
  root = ".",
  read_raw_data = default(read_raw_data),
  read_file_info = default(read_file_info),
  read_method_info = default(read_method_info),
  read_vendor_data_table = default(read_vendor_data_table),
  discard_duplicates = TRUE,
  parallel = FALSE,
  parallel_plan = future::multisession,
  parallel_cores = future::availableCores(),
  cache = default(cache),
  read_cache = default(cache),
  reread_outdated_cache = FALSE,
  quiet = default(quiet),
  cache_files_with_errors = TRUE
)
}
\arguments{
\item{...}{one or multiple file/folder paths. All files must have a supported file extension. All folders are expanded and searched for files with supported file extensions (which are then included in the read).}

\item{root}{root directory for the isofiles. Can be relative to the current working directory (e.g. \code{"data"}) or an absolute path on the file system (e.g. \code{"/Users/..."} or \code{"C:/Data/.."}). The default is the current working directory (\code{"."}). Can be supplied as a vector of same length as the provided paths if the paths have different roots.}

\item{read_raw_data}{whether to read the raw mass/ion data from the file}

\item{read_file_info}{whether to read auxiliary file information (file id, sequence information, etc.)}

\item{read_method_info}{whether to read methods information (standards, processing info)}

\item{read_vendor_data_table}{whether to read the vendor computed data table}

\item{discard_duplicates}{whether to automatically discard files with duplicate file IDs (i.e. duplicate file names). If \code{TRUE} (the default), only the first files are kept and any files with the same file ID are discarded. If \code{FALSE}, all duplicate files are kept but their file IDs are appended with suffix \code{#1}, \code{#2}, etc.}

\item{parallel}{whether to process in parallel based on the number of available CPU cores. This may yield performance increases for files that are slow to parse such as continuous flow isodat files but usually provides little benefit for efficient data formats such as reading from R Data Archives.}

\item{parallel_plan}{which parallel processing strategy to use, see \link[future]{plan}, typically \code{future::multisession} for compatibility with RStudio interactive mode. If supported by the operating system and running in detached mode (not interactively in RStudio) can also use \code{future::multicore}.}

\item{parallel_cores}{how many processor cores to use for parallel processing. By default the maximum available number of cores (\link[future]{availableCores}), which will allow maximal processing speed but may slow other programs running on your machine. Choose a smaller number if you want some processing resources to remain available for other processes. Will issue a warning if too many cores are requested and reset to the maximum available.}

\item{cache}{whether to cache iso_files. Note that R Data Storage files (.rds, see \link{iso_save}) are never cached since they are already essentially in cached form.}

\item{read_cache}{whether to reload from cache if a cached version exists. Note that it will only read from cache if the raw data file has not been modified since. Files that have been modified on disc (e.g. edited in the vendor software) will always be read anew. To automatically reread cached files that were cached by an outdated version of the isoreader package, set the \code{reread_outdated_cache} flag.}

\item{reread_outdated_cache}{whether to re-read outdated cache files whenever they are encountered.}

\item{quiet}{whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages. Set parameter to overwrite global defaults for this function or set global defaults with calls to \link[=iso_info_messages]{iso_turn_info_messages_on} and \link[=iso_info_messages]{iso_turn_info_messages_off}}

\item{cache_files_with_errors}{deprecated. Please use \link{iso_reread_problem_files} instead to selectively re-read all files in a collection of iso files that had been previously read with errors or warnings.}
}
\description{
Load continuous flow data
}
\seealso{
Other isoread functions for different types of IRMS data: 
\code{\link{iso_read_dual_inlet}()},
\code{\link{iso_read_scan}()}
}
\concept{isoread functions for different types of IRMS data}
