% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoread.R
\name{iso_read_orbitrap_sequence}
\alias{iso_read_orbitrap_sequence}
\title{Load orbitrap sequence}
\usage{
iso_read_orbitrap_sequence(
  seq_file_path,
  data_files_path = dirname(seq_file_path),
  ...
)
}
\arguments{
\item{seq_file_path}{the path to the sequence file}

\item{data_files_path}{the path to the folder with the data (default is the same folder as the sequence file)}

\item{...}{additional parameters are passed to \code{\link{iso_read_orbitrap}}}
}
\description{
Load orbitrap data from a sequence file.
}
\seealso{
Other isoread functions for different types of IRMS data: 
\code{\link{iso_read_continuous_flow}()},
\code{\link{iso_read_dual_inlet}()},
\code{\link{iso_read_orbitrap}()},
\code{\link{iso_read_scan}()}
}
\concept{isoread functions for different types of IRMS data}
