% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{iso_set_default_read_parameters}
\alias{iso_set_default_read_parameters}
\title{Set default read options}
\usage{
iso_set_default_read_parameters(
  data = NULL,
  read_raw_data,
  read_file_info,
  read_method_info,
  read_vendor_data_table,
  quiet = default(quiet)
)
}
\arguments{
\item{data}{a data frame - returned invisibly as is if provided (e.g. in the middle of a pipeline)}

\item{read_raw_data}{if provided, set as the default for `read_raw_data` parameters}

\item{read_file_info}{if provided, set as the default for `read_file_info` parameters}

\item{read_method_info}{if provided, set as the default for `read_method_info` parameters}

\item{read_vendor_data_table}{if provided, set as the default for `read_vendor_data_table` parameters}

\item{quiet}{whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages. Set parameter to overwrite global defaults for this function or set global defaults with calls to \link[=iso_info_messages]{iso_turn_info_message_on} and \link[=iso_info_messages]{iso_turn_info_message_off}}
}
\description{
Set default read options
}
\seealso{
Other settings functions: 
\code{\link{iso_caching}},
\code{\link{iso_get_default_reader_parameters}()},
\code{\link{iso_info_messages}}
}
\concept{settings functions}
