% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{iso_show_default_reader_parameters}
\alias{iso_show_default_reader_parameters}
\title{Show the current default parameters}
\usage{
iso_show_default_reader_parameters(data = NULL, func = NULL, ...)
}
\arguments{
\item{data}{a data frame - returned invisibly as is if provided (e.g. in the middle of a pipeline)}

\item{func}{function to use for formatting the reader parameters table}

\item{...}{additional parameters to forward to the \code{func} function}
}
\description{
Shows a table with the default function parameters for this package.
}
\seealso{
Other settings functions: \code{\link{iso_caching}},
  \code{\link{iso_get_default_reader_parameters}},
  \code{\link{iso_info_messages}},
  \code{\link{iso_set_default_reader_parameters}}
}
