% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_binary_files.R
\name{iso_source_file_op_error}
\alias{iso_source_file_op_error}
\title{Throw source file operation error with useful debugging information for the file type.}
\usage{
iso_source_file_op_error(source_file_obj, msg, ...)
}
\arguments{
\item{source_file_obj}{the source file object for which to print the operation error}

\item{msg}{the message to print}

\item{...}{additional parameters depending on source file type}
}
\description{
Throw source file operation error with useful debugging information for the file type.
}
