% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{isoexport_rda}
\alias{isoexport_rda}
\title{Export data to R Data Archive}
\usage{
isoexport_rda(isofiles, filename, folder = ".", quiet = setting("quiet"))
}
\arguments{
\item{isofiles}{collection of isofile objects to export}

\item{filename}{the name of the file to export (without the Rda ending)}

\item{folder}{the folder where to save the Rda file}

\item{quiet}{whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages. Set parameter to overwrite global defaults for this function or set global defaults with calls to \link[=info_messages]{turn_info_message_on} and \link[=info_messages]{turn_info_message_off}}
}
\value{
returns the isofiles object invisibly for use in pipelines
}
\description{
Export data to R Data Archive
}
