% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{isoplot_continuous_flow}
\alias{isoplot_continuous_flow}
\title{Plot chromatogram from continuous flow data}
\usage{
isoplot_continuous_flow(isofiles, masses = NA, ratios = c(),
  time_interval = c(), normalize = FALSE, zoom = NULL,
  panels = c("none", "traces", "files"), colors = c("none", "traces",
  "files"), linetypes = c("none", "traces", "files"))
}
\arguments{
\item{isofiles}{collection of continuous flow isofile objects}

\item{masses}{which masses to plot (e.g. c("45", "44")), NULL (the default) means that all masses will be plotted}

\item{ratios}{which ratios to plot (e.g. c("45/44", "46/44")), not affected by zoom parameter}

\item{time_interval}{which time interval to plot (in the units of time of the iso objects)}

\item{normalize}{whether to normalize all traces (default is FALSE, i.e. no normalization). If TRUE, normalizes each trace across all files. Normalizing always scales such that each trace fills the entire height of the plot area. Note that zooming (if \code{zoom} is set) is applied after normalizing.}

\item{zoom}{if not set, automatically scales to the maximum range in the selected time_interval in each panel. If set, scales by the indicated factor, i.e. values > 1 are zoom in, values < 1 are zoom out, baseline always remains the anchor point. Note that for overlay plots (\code{panels = "none"}) zooming is relative to the max in each panel (potentially across different traces). Also note that zooming only affects masses, ratios are not zoomed.}

\item{panels}{whether to panel traces, options are "none" (overlay all), "traces" (by mass/ratio traces), "files" (panel by files). The default is "traces"}

\item{colors}{whether to color traces, options are the same as for panels but the default is "files"}

\item{linetypes}{whether to differentiate traces by linetype, options are the same as for panels but the default is "none". Note that a limited number of linetypes (6) is defined by default and the plot will fail if a higher number is required unless specified using \code{\link[ggplot2]{scale_linetype}}}
}
\description{
Plot chromatogram from continuous flow data
}
\seealso{
Other plot functions: \code{\link{isoplot_raw_data}}
}
