% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{isoplot_dual_inlet}
\alias{isoplot_dual_inlet}
\title{Plot mass data from dual inlet files}
\usage{
isoplot_dual_inlet(isofiles, masses = NA, ratios = c())
}
\arguments{
\item{isofiles}{collection of isofile objects}

\item{masses}{which masses to plot (e.g. c("45", "44")), NULL (the default) means that all masses will be plotted}

\item{ratios}{which ratios to plot (e.g. c("45/44", "46/44")), not affected by zoom parameter}
}
\description{
Plot mass data from dual inlet files
}
