% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{isoreader-package}
\alias{isoreader}
\alias{isoreader-package}
\title{isoreader: Read IRMS data from isodat files}
\description{
R interface to IRMS (isotope ratio mass spectrometry) file formats typically used in stable isotope geochemistry.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kopflab/isorunner}
  \item Report bugs at \url{https://github.com/kopflab/isorunner/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Kopf \email{sebastian.kopf@colorado.edu}

}
\keyword{internal}
