% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isoreader_example}
\alias{isoreader_example}
\alias{iso_get_reader_examples}
\title{Example files}
\usage{
isoreader_example(filename)

iso_get_reader_examples()
}
\arguments{
\item{filename}{the name of the example file for which to retrieve the system path}
}
\description{
The isoreader package comes with a few example files to make it easy to illustrate the functionality.
}
\details{
\code{isoreader_example}: retrieve the path to an isoreader example file

\code{iso_get_reader_examples}: list of all available isoreader example files
}
