% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_binary_files.R
\name{map_binary_structure}
\alias{map_binary_structure}
\title{Map isodat file binary structure.}
\usage{
map_binary_structure(
  bfile,
  length = 100,
  start = bfile$pos,
  ctrl_blocks = get_ctrl_blocks_config()
)
}
\arguments{
\item{bfile}{the binary file, stored in each iso_file under \code{$binary} if (and only if) the file was read with \link{iso_turn_debug_on} activated before.}

\item{length}{how many bytes to map}

\item{start}{at which byte position to start mapping (index 1 based)}

\item{ctrl_blocks}{named list of block patterns with size, regexp and [optional] replace function}
}
\description{
Map out binary structure for easy visualization (used mostly for error messages and debugging). See the development vignette for details and example application.
}
