% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problems.R
\name{omit_files_with_problems}
\alias{omit_files_with_problems}
\title{Remove problematic files}
\usage{
omit_files_with_problems(isofiles, type = c("error", "warning", "both"),
  quiet = setting("quiet"))
}
\arguments{
\item{isofiles}{collection of isofile objects}

\item{type}{what type of problem causes removal of the file: \code{"error"}, \code{"warning"} or \code{"both"}}

\item{quiet}{whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages. Set parameter to overwrite global defaults for this function or set global defaults with calls to \link[=info_messages]{turn_info_message_on} and \link[=info_messages]{turn_info_message_off}}
}
\description{
Removes the files that have encountered problems, either errors, warnings or both and returns the remaining isofiles. For additional functions available to check for and deal with problems, see the \link{problem_functions}.
}
