% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_dual_inlet}
\alias{plot_dual_inlet}
\title{Plot mass data from dual inlet files}
\usage{
plot_dual_inlet(isofiles, datasets, panel_by = "dataset", color_by = "file",
  linetype_by = "none", shape_by = "SA|STD")
}
\arguments{
\item{isofiles}{collection of isofile objects}

\item{datasets}{which masses and ratios to plot (e.g. c("44", "45", "45/44")), if omitted, all available masses and ratios are plotted. Note that ratios should be calculated using \code{\link{calculate_ratios}} prior to plotting.}

\item{panel_by}{whether to panel data by anything, options are "none" (overlay all), "dataset" (by mass/ratio dataset), "file" (panel by files), "SA|STD" (panel by sample|standard). The default is "dataset"}

\item{color_by}{whether to color data by anything, options are the same as for \code{panel_by} but the default is "file"}

\item{linetype_by}{whether to differentiate data by linetype, options are the same as for \code{panel_by} but the default is "none". Note that a limited number of linetype_by (6) is defined by default and the plot will fail if a higher number is required unless specified using \code{\link[ggplot2]{scale_linetype}}}

\item{shape_by}{whether to shape data points by anything, options are the same as for panel_by but the default is "SA|STD" (sample|standard)}
}
\description{
Plot mass data from dual inlet files
}
\note{
normalization is not useful for dual inlet data, except potentially between standard and sample - however, for this it is more meaningful to simply plot the relevant ratios together
}
