% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_binary_isodat_files.R
\name{print.binary_isodat_file}
\alias{print.binary_isodat_file}
\title{Print formatted isodat structure, first 200 bytes by default}
\usage{
\method{print}{binary_isodat_file}(x, start = x$pos, length = 200, ...)
}
\arguments{
\item{x}{object to show}

\item{start}{where to start to print (set by default by the current position of the file)}

\item{length}{how much to print}

\item{...}{additional parameters (passed to iso_print_source_file_structure)}
}
\description{
Print formatted isodat structure, first 200 bytes by default
}
