% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_binary_files.R
\name{print.binary_structure_map}
\alias{print.binary_structure_map}
\title{Print binary structure map}
\usage{
\method{print}{binary_structure_map}(
  x,
  ...,
  data_as_raw = FALSE,
  line_break_blocks = c("cblock", "stx", "etx"),
  pos_info = TRUE
)
}
\arguments{
\item{x}{object to show.}

\item{...}{additional parameters passed to print.default}

\item{data_as_raw}{whether to show data as raw}

\item{line_break_blocks}{at which blocks to introduce a line break}

\item{pos_info}{whether to include position information}
}
\description{
Print binary structure map
}
