% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isodata_structures.R
\name{print.continuous_flow}
\alias{print.continuous_flow}
\title{Print a continuous_flow isofile}
\usage{
\method{print}{continuous_flow}(x, ..., show_problems = TRUE)
}
\arguments{
\item{x}{Object to show.}

\item{...}{additional parameters passed to print.default}

\item{show_problems}{whether to show encountered problems}
}
\description{
Print a continuous_flow isofile
}
