% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoread.R
\name{print.isofiles}
\alias{print.isofiles}
\title{Print a collection of isofiles}
\usage{
\method{print}{isofiles}(x, ...)
}
\arguments{
\item{x}{Object to show.}

\item{...}{additional parameters passed to print.default}
}
\description{
Print a collection of isofiles
}
