% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoread.R
\name{read_iso_file}
\alias{read_iso_file}
\title{Read individual iso file}
\usage{
read_iso_file(ds, root, path, file_n, files_n, read_from_cache,
  write_to_cache, write_to_cache_if_errors, cachepath, ext, reader_fun,
  reader_options, reader_fun_env)
}
\arguments{
\item{ds}{the basic data structure for the type of iso_file}

\item{root}{root for relative paths. Can be relative to the current working directory (e.g. \code{"data"}) or an absolute path on the file system (e.g. \code{"/Users/..."} or \code{"C:/Data/.."}). The default is the current working directory (\code{"."}). Can be supplied as a vector of same length as the provided paths if the paths have different roots.}

\item{path}{file path}

\item{file_n}{numer of processsed file for info messages}

\item{files_n}{total number of files for info messages}

\item{read_from_cache}{whether to read from cache}

\item{write_to_cache}{whether to write to cache}

\item{write_to_cache_if_errors}{whether to write to cache even if errors are encountered}

\item{cachepath}{path for the cache file}

\item{ext}{file extension}

\item{reader_fun}{file reader function}

\item{reader_options}{list of paramters to be passed on to the reader}

\item{reader_fun_env}{where to find the reader function}
}
\description{
Low level read function for an individual iso file. Usually not called directly but available for methods development.
}
