% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoread.R
\name{reread_files}
\alias{reread_files}
\title{Re-read isofiles}
\usage{
reread_files(isofiles, ..., stop_if_missing = FALSE,
  quiet = setting("quiet"))
}
\arguments{
\item{isofiles}{collection of isofile objects}

\item{...}{additional read parameters that should be used for re-reading the isofiles, see \code{\link{read_dual_inlet}} and \code{\link{read_continuous_flow}} for details}

\item{stop_if_missing}{whether to stop re-reading if any of the original data files are missing (if FALSE, will warn about the missing files but read those that do exist)}

\item{quiet}{whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages. Set parameter to overwrite global defaults for this function or set global defaults with calls to \link[=info_messages]{turn_info_message_on} and \link[=info_messages]{turn_info_message_off}}
}
\description{
Reload the isotope files from their original data files. This is only possible for isofile objects whose file paths still points to the original raw data file.
}
