% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoread.R
\name{reread_isofiles}
\alias{reread_isofiles}
\alias{reread_isofiles_archive}
\title{Re-read isofiles}
\usage{
reread_isofiles(isofiles, ..., stop_if_missing = FALSE,
  quiet = default(quiet))

reread_isofiles_archive(rda_filepaths, ..., stop_if_missing = FALSE,
  quiet = default(quiet))
}
\arguments{
\item{isofiles}{collection of isofile objects}

\item{...}{additional read parameters that should be used for re-reading the isofiles, see \code{\link{read_dual_inlet}} and \code{\link{read_continuous_flow}} for details}

\item{stop_if_missing}{whether to stop re-reading if any of the original data files are missing (if FALSE, will warn about the missing files and keep them unchanged but re-read those that do exist)}

\item{quiet}{whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages. Set parameter to overwrite global defaults for this function or set global defaults with calls to \link[=info_messages]{turn_info_message_on} and \link[=info_messages]{turn_info_message_off}}

\item{rda_filepaths}{the path(s) to the isofiles R data archive(s) to re-read (can be a single file or vector of files)}
}
\description{
Sometimes it is useful to reload isotope files from their original data files (e.g. after upgrading to a newer version of the isoreader package). 
The functions described below are intended to make this very easy. 
However, it is only possible for isofile objects whose file paths still point to the original raw data files.
}
\details{
\code{reread_isofiles} will re-read all the original data files for the passed in \code{isofiles} object. Returns the reread isofile objects.

\code{reread_isofiles_archive} is a convenience function for refreshing saved isofile collections. It will load a specific isofiles R Data Archive (\code{rda_filepath}), re-read all the data from the original data files and save the collection back to the same rda file. The isofiles are returned invisibly.
}
