% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{set_default_read_parameters}
\alias{set_default_read_parameters}
\title{Set default read options}
\usage{
set_default_read_parameters(data, read_raw_data, read_file_info,
  read_method_info, read_vendor_data_table, quiet = default(quiet))
}
\arguments{
\item{data}{a data frame - returned invisibly as is if provided (e.g. in the middle of a pipeline)}

\item{read_raw_data}{if provided, set as the default for `read_raw_data` parameters}

\item{read_file_info}{if provided, set as the default for `read_file_info` parameters}

\item{read_method_info}{if provided, set as the default for `read_method_info` parameters}

\item{read_vendor_data_table}{if provided, set as the default for `read_vendor_data_tabl` parameters}
}
\description{
Set default read options
}
\seealso{
Other settings functions: \code{\link{caching}},
  \code{\link{info_messages}},
  \code{\link{show_isoreader_parameters}}
}
