% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{set_temp}
\alias{set_temp}
\title{Set temporary option}
\usage{
set_temp(name, value)
}
\arguments{
\item{name}{name of the temporary option}

\item{value}{value of the temporary option}
}
\description{
Set a temporary option for parallel processing in isoprocessor.
}
