% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{show_isoreader_parameters}
\alias{show_isoreader_parameters}
\title{Show the current default parameters
Shows a table with the default function parameters for this package.}
\usage{
show_isoreader_parameters(data = NULL)
}
\arguments{
\item{data}{a data frame - returned invisibly as is if provided (e.g. in the middle of a pipeline)}
}
\description{
Show the current default parameters
Shows a table with the default function parameters for this package.
}
\seealso{
Other settings functions: \code{\link{caching}},
  \code{\link{info_messages}},
  \code{\link{set_default_read_parameters}}
}
