% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{vec_arith.iso_double_with_units}
\alias{vec_arith.iso_double_with_units}
\title{vec_arith for iso_double_with_units}
\usage{
\method{vec_arith}{iso_double_with_units}(op, x, y, ...)
}
\description{
vec_arith for iso_double_with_units
}
