% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{vec_cast.iso_double_with_units}
\alias{vec_cast.iso_double_with_units}
\title{vec_cast for iso_double_with_units}
\usage{
\method{vec_cast}{iso_double_with_units}(x, to, ...)
}
\description{
vec_cast for iso_double_with_units
}
