% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{vec_ptype2.iso_double_with_units}
\alias{vec_ptype2.iso_double_with_units}
\title{vec_ptype2 for iso_double_with_units}
\usage{
\method{vec_ptype2}{iso_double_with_units}(x, y, ...)
}
\description{
vec_ptype2 for iso_double_with_units
}
