% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_data.R
\name{aggregate_file_info}
\alias{aggregate_file_info}
\title{Aggregate file info}
\usage{
aggregate_file_info(isofiles, quiet = setting("quiet"))
}
\arguments{
\item{isofiles}{collection of isofile objects}

\item{quiet}{whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages. Set parameter to overwrite global defaults for this function or set global defaults with calls to \link[=info_messages]{turn_info_message_on} and \link[=info_messages]{turn_info_message_off}}
}
\description{
Note file info entries with multiple values are concatenated for this general purpose function.
To get access to a specific multi-value file info, access using $file_info[['INFO_NAME']]
}
\seealso{
Other data retrieval functions: \code{\link{aggregate_raw_data}},
  \code{\link{aggregate_standards_info}},
  \code{\link{aggregate_vendor_data_table}},
  \code{\link{get_file_id}}
}
