% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_data.R
\name{aggregate_standards_info}
\alias{aggregate_standards_info}
\title{Aggregate standards from methods info}
\usage{
aggregate_standards_info(isofiles, with_ratios = TRUE)
}
\arguments{
\item{isofiles}{collection of isofile objects}

\item{with_ratios}{whether to include ratios or just standard delta values}
}
\description{
Aggregates the isotopic standard information recovered from the provided isofiles. Can aggregate just the standards' delta values or combine the delta values with the recovered ratios (if any). Use paramter \code{with_ratios} to exclude/include the ratios.
}
\seealso{
Other data aggregation functions: \code{\link{aggregate_file_info}},
  \code{\link{aggregate_raw_data}},
  \code{\link{aggregate_vendor_data_table}}
}
