% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_data.R
\name{aggregate_vendor_data_table}
\alias{aggregate_vendor_data_table}
\title{Aggregate vendor computed table data}
\usage{
aggregate_vendor_data_table(isofiles, with_units = FALSE,
  select = all_columns(), include_file_info = c(), quiet = default(quiet))
}
\arguments{
\item{isofiles}{collection of isofile objects}

\item{with_units}{whether to include units in the column headers (if there are any) or not (default is FALSE)}

\item{select}{which vendor table columns select. All by default.}

\item{include_file_info}{if provided, will include the requested file information (see \code{\link{aggregate_file_info}}) with the raw data}

\item{quiet}{whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages. Set parameter to overwrite global defaults for this function or set global defaults with calls to \link[=info_messages]{turn_info_message_on} and \link[=info_messages]{turn_info_message_off}}
}
\description{
Aggregate data from the vendor-computed data table. This information is only available if the isofiles were read with parameter \code{read_vendor_data_table=TRUE}.
}
\seealso{
Other data retrieval functions: \code{\link{aggregate_file_info}},
  \code{\link{aggregate_raw_data}},
  \code{\link{aggregate_resistors_info}},
  \code{\link{aggregate_standards_info}},
  \code{\link{get_file_id}}
}
