% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isodata_structures.R
\name{cleanup_isoreader_cache}
\alias{cleanup_isoreader_cache}
\title{Cleanup old cached files}
\usage{
cleanup_isoreader_cache(clean_all = FALSE)
}
\arguments{
\item{clean_all}{if set to TRUE, all cached files will be removed regardless of their version}
}
\description{
Removes cached files generated by isoreader versions different from the currently installed package. Remove all cached files if \code{clean_all} is set to TRUE.
}
