% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isodata_structures.R
\name{is_isofile}
\alias{is_isofile}
\alias{is_isofile_list}
\alias{is_iso_object}
\alias{is_dual_inlet}
\alias{is_continuous_flow}
\alias{as_isofile_list}
\title{Isoreader data structure functions}
\usage{
is_isofile(x)

is_isofile_list(x)

is_iso_object(x)

is_dual_inlet(x)

is_continuous_flow(x)

as_isofile_list(..., discard_duplicates = TRUE)
}
\arguments{
\item{x}{an object to test whether it has the specific class}

\item{...}{isofile and isofile_list objects to concatenate}

\item{discard_duplicates}{whether to discard encountered file id duplicates}
}
\description{
\code{is_isofile} tests if the object is an isofile

\code{is_isofile_list} tests if the object is an isofile list (collection of isofiles)

\code{is_iso_object} test if the object is an iso-object (isofile or isofile list)

\code{is_dual_inlet} tests if an isofile or isofile list consists exclusively of dual inlet file objects

\code{is_continuous_flow} tests if an isofile or isofile list consists exclusively of continuous flow file objects

\code{as_isofile_list} concatenates isofile and isofile list object(s) into one combined isofile list (equivalent to calling \code{c(...)}), flattens all passed lists into one list structure, all individual objects and objects within isofile lists have to be the same type of isofile, issues warnings if there are duplicate file ids and summarizes all problems in the isofile list
}
