% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_to_rda}
\alias{export_to_rda}
\title{Export data to R Data Archive (.rda)}
\usage{
export_to_rda(isofiles, filename, folder = ".", quiet = setting("quiet"))
}
\arguments{
\item{isofiles}{collection of isofile objects}

\item{filename}{the name of the file to export (without the Rda ending)}

\item{folder}{the folder where to save the rda file}

\item{quiet}{whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages. Set parameter to overwrite global defaults for this function or set global defaults with calls to \link[=info_messages]{turn_info_message_on} and \link[=info_messages]{turn_info_message_off}}
}
\value{
returns the isofiles object invisibly for use in pipelines
}
\description{
This function exports the passed in isofiles to an R Data Archive (.rda) file, which is a fairly efficient compressed data storage format. Data exported this way can be easily read back into isoreader using the standard \code{\link{read_continuous_flow}} and \code{\link{read_dual_inlet}} functions or by simply double-clicking the .rda file which will open in an R console and restore the data into variable \code{isofiles}.
}
