% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_data.R
\name{get_file_info}
\alias{get_file_info}
\title{Aggregate file info}
\usage{
get_file_info(isofiles)
}
\arguments{
\item{isofiles}{isotope file objects}
}
\description{
Aggregate file info
}
\seealso{
Other data aggregation functions: \code{\link{get_raw_data}},
  \code{\link{get_vendor_data_table}}
}
