% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_data.R
\name{get_method_info_standards}
\alias{get_method_info_standards}
\title{Get standards from methods info}
\usage{
get_method_info_standards(isofiles, with_ratios = TRUE)
}
\arguments{
\item{isofiles}{isotope file objects}

\item{with_ratios}{whether to include ratios or just standard delta values}
}
\description{
Combines primary and secondary standards into a single table
}
\seealso{
Other data aggregation functions: \code{\link{get_file_info}},
  \code{\link{get_raw_data}},
  \code{\link{get_vendor_data_table}}
}
