% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_data.R
\name{get_vendor_data_table}
\alias{get_vendor_data_table}
\title{Aggregate vendor computed table data}
\usage{
get_vendor_data_table(isofiles, with_units = TRUE)
}
\arguments{
\item{isofiles}{collection of isofile objects}

\item{with_units}{whether to include units in the column headers or not}
}
\description{
Aggregate vendor computed table data
}
\seealso{
Other data aggregation functions: \code{\link{get_file_info}},
  \code{\link{get_method_info_standards}},
  \code{\link{get_raw_data}}
}
