% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoread.R
\name{iso_cleanup_reader_cache}
\alias{iso_cleanup_reader_cache}
\title{Cleanup cached files}
\usage{
iso_cleanup_reader_cache(all = FALSE)
}
\arguments{
\item{all}{deprecated}
}
\description{
Removes all cached files.
}
