% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{iso_export_to_feather}
\alias{iso_export_to_feather}
\title{Export to feather}
\usage{
iso_export_to_feather(iso_files, filepath_prefix,
  include_raw_data = TRUE, include_file_info = TRUE,
  include_method_info = TRUE, include_vendor_data_table = TRUE,
  include_problems = TRUE, quiet = default(quiet))
}
\arguments{
\item{iso_files}{collection of iso_file objects}

\item{filepath_prefix}{the path (folder and filename) prefix for the exported feather files. The correct suffix for different kinds of data and file extension is automatically added}

\item{include_raw_data}{whether to include the raw data in the export (if available)}

\item{include_file_info}{whether to include the file info in the export (if available)}

\item{include_method_info}{whether to include methods infor in the export (if available)}

\item{include_vendor_data_table}{whether to include the vendor data table in the export (if available)}

\item{include_problems}{whether to include the problems table}

\item{quiet}{whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages. Set parameter to overwrite global defaults for this function or set global defaults with calls to \link[=iso_info_messages]{iso_turn_info_message_on} and \link[=iso_info_messages]{iso_turn_info_message_off}}
}
\value{
returns the iso_files object invisibly for use in pipelines
}
\description{
This function exports the passed in iso_files to the Python and R shared feather file format. The different kinds of data (raw data, file info, methods info, etc.) are exported to separate feather files that are saved with the provided \code{filepath_prefix} as prefix. All are only exported if the corresponding \code{include_} parameter is set to \code{TRUE} and only for data types for which this type of data is available and was read (see \code{\link{iso_read_dual_inlet}}, \code{\link{iso_read_continuous_flow}} for details on read parameters). Note that in rare instances where vectorized data columns exist in the file information (e.g. measurement_info), they are concatenated with ', ' in feather output.
}
\seealso{
Other export functions: \code{\link{iso_export_to_excel}},
  \code{\link{iso_export_to_rds}}, \code{\link{iso_save}}
}
\concept{export functions}
