% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_data.R
\name{iso_get_data}
\alias{iso_get_data}
\title{DEPRECATED}
\usage{
iso_get_data(...)
}
\arguments{
\item{...}{forwarded to \link{iso_get_all_data}}
}
\description{
Please use \link{iso_get_all_data} instead.
}
