% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{iso_get_default_reader_parameters}
\alias{iso_get_default_reader_parameters}
\title{Get the current default parameters}
\usage{
iso_get_default_reader_parameters()
}
\description{
Retrieve a table with all default function parameters for this package. 
To set read parameters, see \code{\link{iso_set_default_read_parameters}}. 
To set messaging and caching parameters see \code{\link{iso_info_messages}} and see \code{\link{iso_caching}}.
}
\seealso{
Other settings functions: 
\code{\link{iso_caching}},
\code{\link{iso_info_messages}},
\code{\link{iso_set_default_read_parameters}()}
}
\concept{settings functions}
