% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problems.R
\name{iso_get_problems}
\alias{iso_get_problems}
\title{Retrieve parsing problems}
\usage{
iso_get_problems(iso_files)
}
\arguments{
\item{iso_files}{collection of iso_file objects}
}
\description{
This is identical to the readr \code{\link[readr]{problems}} function.
}
\seealso{
Other problem functions: 
\code{\link{iso_filter_files_with_problems}()},
\code{\link{iso_get_problems_summary}()},
\code{\link{iso_has_problems}()},
\code{\link{iso_problem_functions}}
}
\concept{problem functions}
