% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problems.R
\name{iso_get_problems}
\alias{iso_get_problems}
\title{Retrieve parsing problems}
\usage{
iso_get_problems(iso_files, select = everything())
}
\arguments{
\item{iso_files}{collection of iso_file objects}

\item{select}{which data columns to select - use \code{c(...)} to select multiple, supports all \link[dplyr]{select} syntax. By default, all columns are selected.}
}
\description{
This function retrieves parsing problems encountered during the reading of a set of iso files.
}
\seealso{
Other problem functions: 
\code{\link{iso_filter_files_with_problems}()},
\code{\link{iso_get_problems_summary}()},
\code{\link{iso_has_problems}()},
\code{\link{iso_problem_functions}}
}
\concept{problem functions}
