% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_data.R
\name{iso_get_standards_info}
\alias{iso_get_standards_info}
\title{DEPRECATED}
\usage{
iso_get_standards_info(...)
}
\arguments{
\item{...}{forwarded to \link{iso_get_standards}}
}
\description{
Please use \link{iso_get_standards} instead.
}
