% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoread.R
\name{iso_get_supported_file_types}
\alias{iso_get_supported_file_types}
\title{Supported file types}
\usage{
iso_get_supported_file_types()
}
\description{
Get an overview of all the file types currently supported by the isoreader package. To register additional file readers, use the \code{\link{iso_register_dual_inlet_file_reader}} and \code{\link{iso_register_continuous_flow_file_reader}} functions.
}
\seealso{
Other file_types: 
\code{\link{iso_register_dual_inlet_file_reader}()}
}
\concept{file_types}
