% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_data.R
\name{iso_get_vendor_data_table}
\alias{iso_get_vendor_data_table}
\title{Aggregate vendor computed table data}
\usage{
iso_get_vendor_data_table(iso_files, with_units = FALSE,
  select = all_columns(), include_file_info = c(), quiet = default(quiet))
}
\arguments{
\item{iso_files}{collection of iso_file objects}

\item{with_units}{whether to include units in the column headers (if there are any) or not (default is FALSE)}

\item{select}{which vendor table columns select. All by default.}

\item{include_file_info}{if provided, will include the requested file information (see \code{\link{iso_get_file_info}}) with the raw data}

\item{quiet}{whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages. Set parameter to overwrite global defaults for this function or set global defaults with calls to \link[=iso_info_messages]{turn_info_message_on} and \link[=iso_info_messages]{turn_info_message_off}}
}
\description{
Aggregate data from the vendor-computed data table. This information is only available if the iso_files were read with parameter \code{read_vendor_data_table=TRUE}.
}
\seealso{
Other data retrieval functions: \code{\link{iso_get_file_info}},
  \code{\link{iso_get_raw_data}},
  \code{\link{iso_get_resistors_info}},
  \code{\link{iso_get_standards_info}}
}
