% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problems.R
\name{iso_omit_files_with_problems}
\alias{iso_omit_files_with_problems}
\title{Renamed to iso_filter_files_with_problems}
\usage{
iso_omit_files_with_problems(...)
}
\arguments{
\item{...}{deprecated}
}
\description{
This function has been renamed to \link{iso_filter_files_with_problems} for naming consistency.
}
