% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{iso_plot_continuous_flow_data}
\alias{iso_plot_continuous_flow_data}
\title{moved to isoprocessor}
\usage{
iso_plot_continuous_flow_data(...)
}
\arguments{
\item{...}{deprecated}
}
\description{
moved to isoprocessor
}
