% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoread.R
\name{iso_read_dual_inlet}
\alias{iso_read_dual_inlet}
\title{Load dual inlet data}
\usage{
iso_read_dual_inlet(..., read_raw_data = default(read_raw_data),
  read_file_info = default(read_file_info),
  read_method_info = default(read_method_info),
  read_vendor_data_table = default(read_vendor_data_table),
  discard_duplicates = TRUE, parallel = TRUE, cache = default(cache),
  read_cache = default(cache), quiet = default(quiet))
}
\arguments{
\item{...}{one or multiple file/folder paths. All files must have a supported file extension. All folders are expanded and searched for files with supported file extensions (which are then included in the read).}

\item{read_raw_data}{whether to read the raw mass/ion data from the file}

\item{read_file_info}{whether to read auxiliary file information (file id, sequence information, etc.)}

\item{read_method_info}{whether to read methods information (standards, processing info)}

\item{read_vendor_data_table}{whether to read the vendor computed data table}

\item{discard_duplicates}{whether to automatically discard files with duplicate file IDs (i.e. duplicate file names). If \code{TRUE} (the default), only the first files are kept and any files with the same file ID are discarded. If \code{FALSE}, all duplicate files are kept but their file IDs are appended with suffix \code{#1}, \code{#2}, etc.}

\item{parallel}{whether to process in parallel based on the number of available CPU cores}

\item{cache}{whether to cache iso_files. Note that previously exported R Data Archives (di.rda, cf.rda) are never cached since they are already essentially in cached form.}

\item{read_cache}{whether to reload from cache if a cached version exists. Note that it will only read from cache if the file was previously read with the exact same isoreader version and read options and has not been modified since.}

\item{quiet}{whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages. Set parameter to overwrite global defaults for this function or set global defaults with calls to \link[=iso_info_messages]{iso_turn_info_message_on} and \link[=iso_info_messages]{iso_turn_info_message_off}}
}
\description{
Load dual inlet data
}
\seealso{
Other isoread functions for different types of IRMS data: \code{\link{iso_read_continuous_flow}}
}
\concept{isoread functions for different types of IRMS data}
