% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoread.R
\name{iso_read_files}
\alias{iso_read_files}
\title{Core function to read isotope data files}
\usage{
iso_read_files(paths, supported_extensions, data_structure, ...,
  discard_duplicates = TRUE, cache = default(cache),
  read_cache = default(cache), quiet = default(quiet))
}
\arguments{
\item{paths}{one or multiple file/folder paths. All files must have a supported file extension. All folders are expanded and searched for files with supported file extensions (which are then included in the read).}

\item{supported_extensions}{data frame with supported extensions and corresponding reader functions}

\item{data_structure}{the basic data structure for the type of iso_file}

\item{...}{additional parameters passed to the specific processing functions for the different file extensions}

\item{discard_duplicates}{whether to automatically discard files with duplicate file IDs (i.e. duplicate file names). If \code{TRUE} (the default), only the first files are kept and any files with the same file ID are discarded. If \code{FALSE}, all duplicate files are kept but their file IDs are appended with suffix \code{#1}, \code{#2}, etc.}

\item{cache}{whether to cache iso_files. Note that previously exported R Data Archives (di.rda, cf.rda) are never cached since they are already essentially in cached form.}

\item{read_cache}{whether to reload from cache if a cached version exists. Note that it will only read from cache if the file was previously read with the exact same isoreader version and read options and has not been modified since.}

\item{quiet}{whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages. Set parameter to overwrite global defaults for this function or set global defaults with calls to \link[=iso_info_messages]{iso_turn_info_message_on} and \link[=iso_info_messages]{iso_turn_info_message_off}}
}
\value{
single iso_file object (if single file) or list of iso_files (iso_file_list)
}
\description{
This function takes care of extracting basic information about iso_files, dealing with problems and making sure only valid fire formats are processed. 
This function is not typicaly called directly but indirectly by calling \link{iso_read_dual_inlet}, \link{iso_read_continuous_flow} and \link{iso_read_scan}. 
It is made available outside the package because it can be very useful for testing new file readers.
}
