% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{isoplot_raw_data}
\alias{isoplot_raw_data}
\title{Plot raw data from isoreader files}
\usage{
isoplot_raw_data(isofiles, ...)
}
\arguments{
\item{isofiles}{collection of isofile objects to plot}

\item{...}{parameters for the data specific plotting functions}
}
\description{
Convenience function for making standard plots for raw isoreader data. Calls \link{\code{isoplot_continuous_flow}} and \link{\code{isoplot_dual_inlet}} for data specific plotting (see those functions for parameter details).
}
\seealso{
Other plot functions: \code{\link{isoplot_continuous_flow}}
}
