% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{isoread}
\alias{isoread}
\title{read isotope data file}
\usage{
isoread(...)
}
\arguments{
\item{...}{original isoread parameters}
}
\description{
Deprecated, use [isoread_dual_inlet()], [isoread_continuous_flow()] and [isoread_scan()] instead.
}
