% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoread.R
\name{isoread}
\alias{isoread}
\title{Read isotope data file}
\usage{
isoread(...)
}
\arguments{
\item{...}{original isoread parameters}
}
\description{
Deprecated, use \link{read_dual_inlet}, \link{read_continuous_flow} and \link{read_scan} instead.
}
