% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoread.R
\name{isoread}
\alias{isoread}
\title{Read isotope data file}
\usage{
isoread(...)
}
\arguments{
\item{...}{original isoread parameters}
}
\description{
This function from the original isoread package is deprecated, please use \link{iso_read_dual_inlet}, \link{iso_read_continuous_flow} and \link{iso_read_scan} instead.
}
