% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual_inlet.R
\name{isoread_dual_inlet}
\alias{isoread_dual_inlet}
\title{Load dual inlet data}
\usage{
isoread_dual_inlet(paths, read_raw_data = TRUE, read_file_info = TRUE,
  read_method_info = TRUE, read_vendor_data_table = FALSE,
  quiet = setting("quiet"), cache = setting("cache"))
}
\arguments{
\item{paths}{one or multiple file/folder paths. All files must have a supported file extension. All folders are expanded and searched for files with supported file extensions (which are then included in the read).}

\item{read_raw_data}{whether to read the raw mass/ion data from the file}

\item{read_file_info}{whether to read auxiliary file information (file id, sequence information, etc.)}

\item{read_method_info}{whether to read methods information (standards, processing info)}

\item{read_vendor_data_table}{whether to read the vendor computed data table}

\item{quiet}{whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages. Set parameter to overwrite global defaults for this function or set global defaults with calls to \link[=info_messages]{turn_info_message_on} and \link[=info_messages]{turn_info_message_off}}

\item{cache}{whether to cache isofiles and attempt to reload from cache (will only reload if a file was previously read with the same read options and has NOT been modified since)}
}
\description{
Load dual inlet data
}
\seealso{
Other isoread functions for different types of IRMS data: \code{\link{isoread_continuous_flow}}
}
