% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual_inlet.R
\name{isoread_dual_inlet}
\alias{isoread_dual_inlet}
\title{Load dual inlet data}
\usage{
isoread_dual_inlet(paths, quiet = default("quiet"))
}
\arguments{
\item{quiet}{whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages. Set parameter to overwrite global defaults for this function or set global defaults with calls to \link[=info_messages]{turn_info_message_on} and \link[=info_messages]{turn_info_message_off}}
}
\description{
Load dual inlet data
}
