% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual_inlet.R
\name{isoread_dual_inlet}
\alias{isoread_dual_inlet}
\title{Load dual inlet data}
\usage{
isoread_dual_inlet(paths, quiet = setting("quiet"), read_mass_data = TRUE,
  read_file_info = TRUE)
}
\arguments{
\item{paths}{one or multiple file/folder paths. All files must have a supported file extension. All folders are expanded and searched for files with supported file extensions (which are then included in the read).}

\item{quiet}{whether to display (quiet=FALSE) or silence (quiet = TRUE) information messages. Set parameter to overwrite global defaults for this function or set global defaults with calls to \link[=info_messages]{turn_info_message_on} and \link[=info_messages]{turn_info_message_off}}

\item{read_mass_data}{whether to read the raw mass data from the file}

\item{read_file_info}{whether to read auxiliary file information (program, methods, etc.)}
}
\description{
Load dual inlet data
}
