% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{isoreader-package}
\alias{isoreader}
\alias{isoreader-package}
\title{isoreader: Read Stable Isotope Data Files}
\description{
Interface to the raw data file formats commonly encountered in scientific disciplines that make use of stable isotopes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/isoverse/isoreader}
  \item Report bugs at \url{https://github.com/isoverse/isoreader/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Kopf \email{sebastian.kopf@colorado.edu} (\href{https://orcid.org/0000-0002-2044-0201}{ORCID})

Authors:
\itemize{
  \item Brett Davidheiser-Kroll \email{brett.davidheiserkroll@colorado.edu} (\href{https://orcid.org/0000-0002-6153-7851}{ORCID})
  \item Ilja Kocken \email{i.j.kocken@uu.nl} (\href{https://orcid.org/0000-0003-2196-8718}{ORCID})
}

}
\keyword{internal}
