% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R, R/package.R, R/problems.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{parse_number}
\alias{parse_logical}
\alias{parse_integer}
\alias{parse_double}
\alias{parse_datetime}
\alias{!!}
\alias{!!!}
\alias{\%>\%}
\alias{everything}
\alias{starts_with}
\alias{ends_with}
\alias{matches}
\alias{filter}
\alias{tibble}
\alias{problems}
\alias{stop_for_problems}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{filter}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{readr}{\code{\link[readr]{parse_datetime}}, \code{\link[readr]{parse_double}}, \code{\link[readr]{parse_integer}}, \code{\link[readr]{parse_logical}}, \code{\link[readr]{parse_number}}, \code{\link[readr]{problems}}, \code{\link[readr]{stop_for_problems}}}

  \item{rlang}{\code{\link[rlang]{!!}}, \code{\link[rlang]{!!!}}}

  \item{tibble}{\code{\link[tibble]{tibble}}}

  \item{tidyselect}{\code{\link[tidyselect]{ends_with}}, \code{\link[tidyselect]{everything}}, \code{\link[tidyselect]{matches}}, \code{\link[tidyselect]{starts_with}}}
}}

