% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problems.R, R/utils.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{problems}
\alias{reexports}
\alias{stop_for_problems}
\alias{reexports}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{readr}{\code{\link[readr]{problems}}, \code{\link[readr]{stop_for_problems}}}
}}

