% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R, R/problems.R, R/utils.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{parse_number}
\alias{reexports}
\alias{parse_logical}
\alias{reexports}
\alias{parse_integer}
\alias{reexports}
\alias{parse_double}
\alias{reexports}
\alias{stop_for_problems}
\alias{reexports}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{readr}{\code{\link[readr]{parse_number}}, \code{\link[readr]{parse_logical}}, \code{\link[readr]{parse_integer}}, \code{\link[readr]{parse_double}}, \code{\link[readr]{stop_for_problems}}}
}}

