% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{vec_arith.iso_double_with_units}
\alias{vec_arith.iso_double_with_units}
\title{vec_arith for iso_double_with_units}
\usage{
\method{vec_arith}{iso_double_with_units}(op, x, y, ...)
}
\arguments{
\item{op}{An arithmetic operator as a string}

\item{x}{A pair of vectors. For \code{!}, unary \code{+} and unary \code{-}, \code{y} will be
a sentinel object of class \code{MISSING}, as created by \code{MISSING()}.}

\item{y}{A pair of vectors. For \code{!}, unary \code{+} and unary \code{-}, \code{y} will be
a sentinel object of class \code{MISSING}, as created by \code{MISSING()}.}

\item{...}{These dots are for future extensions and must be empty.}
}
\description{
vec_arith for iso_double_with_units
}
