% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{vec_cast.iso_double_with_units}
\alias{vec_cast.iso_double_with_units}
\title{vec_cast for iso_double_with_units}
\usage{
\method{vec_cast}{iso_double_with_units}(x, to, ...)
}
\arguments{
\item{x}{Vectors to cast.}

\item{to}{Type to cast to. If \code{NULL}, \code{x} will be returned as is.}

\item{...}{For \code{vec_cast_common()}, vectors to cast. For
\code{vec_cast()}, \code{vec_cast_default()}, and \code{vec_restore()}, these
dots are only for future extensions and should be empty.}
}
\description{
vec_cast for iso_double_with_units
}
